% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extra_tools.R
\name{biv_dens}
\alias{biv_dens}
\title{Estimate the joint density of two random variables (markers)}
\usage{
biv_dens(dat, ngrids = 512, maxs = "default")
}
\arguments{
\item{dat}{is a matrix of intensity of two random variables.}

\item{ngrids}{is the number of grid-points to be used in estimating the kernel density. The diagonal
multivariate plug-in bandwidth estimator is used.}

\item{maxs}{is either a character named "default" or a vector of two values which will respectively be
the maximum values considered for the two variables. If maxs = "default", the maximum of each of the data
vectors are considered.}
}
\value{
It returns a kde object with estimated density and the grid-points from the ks package.
}
\description{
Estimate the joint density of two random variables (markers)
}
