% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extra_tools.R
\name{biv_dens_plot}
\alias{biv_dens_plot}
\title{Plot the estimated joint density of two random variables (markers)}
\usage{
biv_dens_plot(fit, maxs = c(0.5, 0.5))
}
\arguments{
\item{fit}{is the output of the function biv_dens.}
}
\value{
Plots the estimated bivariate joint density.
}
\description{
Plot the estimated joint density of two random variables (markers)
}
