% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survival.R
\name{Cox_PH}
\alias{Cox_PH}
\title{Fit CoxPH model to study association between survival/recurrence outcome and estimated EQMi*}
\usage{
Cox_PH(clinical_dat, cov, EQMI_dat, degree = 1)
}
\arguments{
\item{clinical_dat}{is a data.frame whose first column is "ID" denoting the subject IDs. The next two columns
respectively correspond to the time (time to death or time to recurrence) and censoring indicator
( = 0 imples death or recurrence).}

\item{cov}{is a data.frame whose first column is "ID" denoting the subject IDs. The other columns correspond
to different covariates.}

\item{EQMI_dat}{corresponds to the output of the function QMI_of_all_images. It is a matrix of estimated
EQMI* of the sets of markers, (1, 2), (1, 2, 3), (1, 2, 3, 4), ..., and (1, 2, ..., p), where p is the
number of available markers.}

\item{degree}{corresponds to the degree of polynomial to use in the CoxPH model. If degree > 1,
a non-linear relationship between the clinical outcomes and EQMI* is tested.}
}
\value{
It returns the p-values of the association tests between the clinical outcome and EQMI* of
the different sets of markers.
}
\description{
Fit CoxPH model to study association between survival/recurrence outcome and estimated EQMi*
}
