% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizing_functions.R
\name{dnorm_mine}
\alias{dnorm_mine}
\title{Compute the PDF of a univariate normal distribution.}
\usage{
dnorm_mine(x, mean, sd)
}
\arguments{
\item{x}{is the point where the density will be evaluated at.}

\item{mean}{is the mean of the normal distribution.}

\item{sd}{is the standard deviation of the normal distribution.}
}
\value{
It returns the density  of a univariate normal distribution with specified mean and sd at point x.
}
\description{
Compute the PDF of a univariate normal distribution.
}
