% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marker_data}
\alias{marker_data}
\title{Marker intensity data.}
\format{
A data frame with 190276 rows and 6 variables:
\describe{
  \item{ID}{Subject IDs}
  \item{HLA.DR}{Scaled intensity of HLA-DR marker}
  \item{CD45RO}{Scaled intensity of CD45RO marker}
  \item{HLA_Class_1}{Scaled intensity of HLA_Class_1 marker}
  \item{H3K27me3}{Scaled intensity of H3K27me3 marker}
  \item{H3K9ac}{Scaled intensity of H3K9ac marker}
}
}
\source{
\url{https://www.angelolab.com/mibi-data}
}
\usage{
marker_data
}
\description{
A dataset with expression data of five markers, HLA-DR, CD45RO, H3K27me3, H3K9ac
and HLA_Class_1 in different cells of 39 subjects.
}
\keyword{datasets}
