% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPP.R
\name{NPP}
\alias{NPP}
\title{Net primary productivity (NPP) estimation from temperature and precipitation data}
\usage{
NPP(temp, pre)
}
\arguments{
\item{temp}{temperture precipitation vector}

\item{pre}{precipitation vector}
}
\description{
This function allows you to estimate NPP
from average annual temperture and cummulative precpitation data.
}
\keyword{NPP}
