% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPPr.R
\name{NPPp}
\alias{NPPp}
\title{Net primary productivity (NPP) estimation from precipitation data}
\usage{
NPPp(pre)
}
\arguments{
\item{pre}{precipitation vector}
}
\description{
This function allows you to estimate NPP from  cummulative precpitation data.
}
\keyword{NPP}
