% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPPt.R
\name{NPPt}
\alias{NPPt}
\title{Net primary productivity (NPP) estimation from temperature data}
\usage{
NPPt(temp)
}
\arguments{
\item{temp}{temperture precipitation vector}
}
\description{
This function allows you to estimate NPP from average annual temperture.
}
\keyword{NPP}
