% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMI.R
\name{QMI}
\alias{QMI}
\title{Compute different mutual information (MI) theoretic measures such as EQMI*, EQMI and CSQMI
in a general dataset}
\usage{
QMI(dat, bandwidth = "HPI", measure = "EQMI_star", var_names = F)
}
\arguments{
\item{dat}{is the data matrix whose p columns correspond to p different random variables (markers in a
multiplex imaging dataset) and whose rows are samples (usually, different cells from different subjects in a
multiplex imaging dataset).}

\item{bandwidth}{is the bandwidth selection procedure user wants to use. The default option "HPI"
corresponds to multivariate plug-in bandwidth estimates. For faster computation, use "Ind" which would
independently select bandwidth for every r.v. by using Silverman's rule.}

\item{measure}{corresponds to the MI measure one wants to compute. For example, measure = "EQMI_star"
will output the measure EQMI_star. Other options are, "EQMI, "CSQMI" and "ALL", where the last one
outputs all the three measures.}

\item{var_names}{can be \code{TRUE} (T) or \code{FALSE} (F). If it is True the output list will have name of the
measures.}
}
\value{
It returns a list whose first element is the vector of estimated MI between the following combination
of the markers, (1, 2), (1, 2, 3), (1, 2, 3, 4), ..., and (1, 2, ..., p), and the second element is the
vector of bandiwidth parameters used. If measure = "All", the first element is a list of three elements
corresponding to \code{EQMI_star}, \code{EQMI} and \code{CSQMI}.
}
\description{
Compute different mutual information (MI) theoretic measures such as EQMI*, EQMI and CSQMI
in a general dataset
}
