% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMI_of_all_images.R
\name{QMI_all}
\alias{QMI_all}
\title{Compute different mutual information (MI) theoretic measures such as EQMI*, EQMI and CSQMI
for every image (subject) in a multiplex imaging dataset}
\usage{
QMI_all(data, bandwidth = "HPI", measure = "EQMI_star", progress_bar = "True")
}
\arguments{
\item{data}{is the data.frame of marker intensity observed in different cells of different subjects.
The first column is "ID" which corresponds to the subject ID i.e., the subject to which the row
(cell) belongs to. The rest of the p columns correspond to p available markers.}

\item{bandwidth}{is the bandwidth selection procedure user wants to use. The default option "HPI"
corresponds to multivariate plug-in bandwidth estimates. For faster computation, use "Ind" which would
independently select bandwidth for every r.v. by using Silverman's rule.}

\item{measure}{corresponds to the MI measure one wants to compute. For example, measure = \code{EQMI_star}
will output the measure EQMI_star. Other options are, \code{EQMI} and \code{CSQMI}. We recommend using
\code{EQMI_star} for association analysis with clinical outcomes down the line.}

\item{progress_bar}{if \code{TRUE} will show a progress bar.}
}
\value{
It returns a data.frame whose first column is the vector of subject ID's. The other (p-1) columns
respectively correspond to estimated MI between the sets of markers, (1, 2), (1, 2, 3), (1, 2, 3, 4), ..., and
(1, 2, ..., p).
}
\description{
Compute different mutual information (MI) theoretic measures such as EQMI*, EQMI and CSQMI
for every image (subject) in a multiplex imaging dataset
}
