% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizing_functions.R
\name{range01}
\alias{range01}
\title{Standardizing a marker expression vector between 0 and 1.}
\usage{
range01(x, ...)
}
\arguments{
\item{x}{is the vector of expression of a marker.}
}
\value{
It returns a standardized expression vector.
}
\description{
Standardizing a marker expression vector between 0 and 1.
}
