% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extra_tools.R
\name{univ_dens}
\alias{univ_dens}
\title{Estimate the marginal density of a marker (random variable)}
\usage{
univ_dens(dat, ngrids = 1024)
}
\arguments{
\item{dat}{is a vector with intensity data of a single marker or a matrix of intensity of multiple markers.}

\item{ngrids}{number of grid-points to be used in estimating the kernel density. The bandwidth is chosen
using Silverman's rule.}
}
\value{
It returns a list whose first element is the estimated density and the second element is the grid-
points where the density is estimated.
}
\description{
Estimate the marginal density of a marker (random variable)
}
