% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extra_tools.R
\name{univ_dens_plot}
\alias{univ_dens_plot}
\title{Plot the estimated marginal density of a random variable (marker)}
\usage{
univ_dens_plot(f)
}
\arguments{
\item{f}{is the output of the function univ_dens.}
}
\value{
Plots the estimated marginal density.
}
\description{
Plot the estimated marginal density of a random variable (marker)
}
